Made materials change in real time when compared to maya.
-- Implementation details --
Added one individual texture and color material for each mesh kept track of via maps in the mayaviewer.
In addition each time a material changes properties and is connected to a mesh it is sent as a "material switch" to each connected mesh in order for the viewer to render it.

I would prefer to implement this another way but gameplay3D does not appear to support multiple meshes using the same material.