Changed Comlib in order to work properly. It no longer has an infinite while loop for the consumer if it catches up but rather just returns false.
This also allows for movement and resizing of the window. I also ensured the viewer functions with this change and removed the faulty solution. The camera now pans in real time.

However once a very high amount of messages are sent. Something causes the masterheader to missread and causes crash in the viewer. Cause as of yet unknown.