Renaming entirely functional.
New known bug: Objects in seperate hierarchies with the same name is not yet handled as all names are assumed to be unique.

Potential future development would most likely benefit from a name overhaul. Specifically when it comes to hierarchies but also within sending messages as making all names dynamic in size would be preferable.